+++ src/ch17-03-more-futures.md
@@ -156,7 +156,7 @@ little involved, so let’s walk through each part of it.
 <Listing number="17-TODO" caption="" file-name="src/main.rs">
 
 ```rust,ignore,does_not_compile
-{{#rustdoc_include ../listings/ch17-async-await/listing-17-12/src/main.rs:here}}
+{{#rustdoc_include ../listings/ch17-async-await/listing-17-12-orig/src/main.rs:here}}
 ```
 
 </Listing>
